Updates of this patch:
I) gadd now use batch insert for rdf file load with high performance.
It is notebal that this update opertation is not transactional.
II) gsub now use batch remove in the same manner of gadd described in I)
III) fix some bugs of transaction processing and add a stressing testing for it